requirements.txt
Flask==2.0.1
Flask-MySQLdb==0.2.0
bcrypt==3.2.0
cryptography==41.0.2
python-dotenv==0.19.0
werkzeug==2.0.3
pyotp==2.6.0
qrcode[pil]==7.3.1